refactor(issuance): extract usd calculation
Sending the USD price to the DB only to multiply it is silly. The query
makes sense only if the USD price is available in the DB already but it
isn't on a slot by slot basis so we multiply by an average. If we
multiple by an average anyway, we may as well do that server side, no
need to make the DB do extra work.