Added a tool to detect and fix corrupted blob burn data.
I've tested the frontend unit conversion first but didn't find any bugs,
and since there have been no recent changes to either the frontend or
backend I suspect some external influence corrupted a block in the
database, possibly caused by yesterday's cloudflare outage as the
timing seems to line up.

This tool scans the database and looks for blocks with impossible values
and reports back any findings. The tool can also be used to remove those
blocks from the database, when the application is restarted it should
repopulate the database with the correct values automatically.

usage:
  cargo run --bin fix-blob-corruption -- --mode=diagnose
  cargo run --bin fix-blob-corruption -- --mode=fix --confirm